backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
gunicorn==22.0.0

faiss-cpu==1.8.0
sentence-transformers==3.0.1
torch==2.2.2

numpy==1.26.4
scikit-learn==1.5.2

requests==2.32.3
python-dotenv==1.0.1
gdown==5.2.0